requirements.txt
slack_sdk
flask
openai
pinecone
python-dotenv 